Version 1.0.32.0 - 100G: adding support for GNR-D NIC (NS) Advanced 2XNAC mode base - 100G: fix for compilation issue on SLES 12.5 - generic: fix for not showing any adapter and exiting with error code 4 - generic: fixes for Coverity hits - generic: adding package_file.c and package_file.h to the repository

+++ devlink_module/src/qdl_msg.c
@@ -1243,7 +1243,7 @@ uint8_t* _qdl_put_msg_header(uint8_t *msg, uint16_t type, uint16_t flags)
 	header->nlmsg_len = NLA_ALIGN(sizeof(*header));
 	header->nlmsg_type = type;
 	header->nlmsg_flags = flags;
-	header->nlmsg_seq = time(NULL);
+	header->nlmsg_seq = (uint64_t)time(NULL);		/* time_t is 64-bit wide from GCC 11, casting fixes Coverity hits */
 	header->nlmsg_pid = 0;
 
 	return msg;
